lab 17 part 2

+++ 17_servlets_jsp/servlets_task2/src/main/java/com/nixsolutions/dao/AuthorBookDao.java
@@ -6,9 +6,9 @@ import com.nixsolutions.entity.AuthorBook;
 
 public interface AuthorBookDao {
 	public List<AuthorBook> getAllAuthorBook();
-	public List<AuthorBook> getBooksIdByAuthorId(int authorId);
-	public List<AuthorBook> getAuthorIdByBookId(int bookId);
-	public AuthorBook getAuthorBookById(int authorId, int bookId);
+	public List<AuthorBook> getBooksIdByAuthorId(Long authorId);
+	public List<AuthorBook> getAuthorIdByBookId(Long bookId);
+	public AuthorBook getAuthorBookById(Long authorId, Long bookId);
 	public void createAuthorBook(AuthorBook authorBook);
 	public void updateAuthorBook(AuthorBook authorBook);
 	public void deleteAuthorBook(AuthorBook authorBook);
